# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Pace Running Tracker API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type RunType = typeof RunType[keyof typeof RunType];


export const RunType = {
  Easy: 'Easy',
  Tempo: 'Tempo',
  Long: 'Long',
  Recovery: 'Recovery',
  Interval: 'Interval',
} as const;

export interface Run {
  id: number;
  /** @nullable */
  name?: string | null;
  type: RunType;
  /** Distance in km */
  distance: number;
  /** Duration in seconds */
  duration: number;
  /** Pace in seconds per km */
  pace: number;
  /** @nullable */
  calories?: number | null;
  /**
     * Elevation gain in meters (placeholder)
     * @nullable
     */
  elevation?: number | null;
  /** ISO datetime of run */
  date: string;
  /**
     * JSON array of [lat, lng] coordinates
     * @nullable
     */
  route?: string | null;
  /**
     * JSON array of per-km split times in seconds
     * @nullable
     */
  splits?: string | null;
  createdAt: string;
}

export type RunInputType = typeof RunInputType[keyof typeof RunInputType];


export const RunInputType = {
  Easy: 'Easy',
  Tempo: 'Tempo',
  Long: 'Long',
  Recovery: 'Recovery',
  Interval: 'Interval',
} as const;

export interface RunInput {
  name?: string;
  type: RunInputType;
  distance: number;
  duration: number;
  calories?: number;
  elevation?: number;
  date: string;
  /** JSON array of [lat, lng] coordinates */
  route?: string;
  splits?: string;
}

export type RunUpdateType = typeof RunUpdateType[keyof typeof RunUpdateType];


export const RunUpdateType = {
  Easy: 'Easy',
  Tempo: 'Tempo',
  Long: 'Long',
  Recovery: 'Recovery',
  Interval: 'Interval',
} as const;

export interface RunUpdate {
  name?: string;
  type?: RunUpdateType;
  distance?: number;
  duration?: number;
  calories?: number;
  elevation?: number;
  date?: string;
  route?: string;
  splits?: string;
}

export interface WeekStats {
  distance: number;
  duration: number;
  runs: number;
  elevation: number;
}

export interface AllTimeStats {
  distance: number;
  duration: number;
  runs: number;
}

export interface WeekProgress {
  completed: number;
  goal: number;
  percentage: number;
}

export interface PersonalRecords {
  /**
     * Pace in seconds per km
     * @nullable
     */
  fastest5k?: number | null;
  /** @nullable */
  fastest10k?: number | null;
  /**
     * Distance in km
     * @nullable
     */
  longestRun?: number | null;
  /**
     * Duration in seconds
     * @nullable
     */
  longestDuration?: number | null;
}

export interface DayStatus {
  date: string;
  dayLabel: string;
  hasRun: boolean;
  isToday: boolean;
}

export interface StatsSummary {
  thisWeek: WeekStats;
  lastWeek: WeekStats;
  allTime: AllTimeStats;
  weekProgress: WeekProgress;
  personalRecords: PersonalRecords;
  weeklyRunDays: DayStatus[];
}

export interface HistoryPoint {
  date: string;
  distance: number;
  runs: number;
}

export interface DayOfWeekCount {
  day: string;
  count: number;
}

export interface StatsHistory {
  points: HistoryPoint[];
  runsPerDayOfWeek: DayOfWeekCount[];
}

export interface ErrorResponse {
  error: string;
}

export type ListRunsParams = {
/**
 * Filter by name or type
 */
search?: string;
/**
 * Max results
 */
limit?: number;
/**
 * Offset for pagination
 */
offset?: number;
};

export type GetRecentRunsParams = {
/**
 * Number of recent runs to return
 */
limit?: number;
};

export type GetStatsHistoryParams = {
/**
 * Time range
 */
range?: GetStatsHistoryRange;
};

export type GetStatsHistoryRange = typeof GetStatsHistoryRange[keyof typeof GetStatsHistoryRange];


export const GetStatsHistoryRange = {
  '7d': '7d',
  '30d': '30d',
  '3m': '3m',
  '6m': '6m',
  '1y': '1y',
  all: 'all',
} as const;