# lib/api-zod/src/generated/types/dayStatus.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Pace Running Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface DayStatus {
  date: string;
  dayLabel: string;
  hasRun: boolean;
  isToday: boolean;
}