# lib/api-zod/src/generated/types/historyPoint.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Pace Running Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface HistoryPoint {
  date: string;
  distance: number;
  runs: number;
}