# lib/api-zod/src/generated/types/run.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Pace Running Tracker API
 * OpenAPI spec version: 0.1.0
 */
import type { RunType } from './runType';

export interface Run {
  id: number;
  /** @nullable */
  name?: string | null;
  type: RunType;
  /** Distance in km */
  distance: number;
  /** Duration in seconds */
  duration: number;
  /** Pace in seconds per km */
  pace: number;
  /** @nullable */
  calories?: number | null;
  /**
     * Elevation gain in meters (placeholder)
     * @nullable
     */
  elevation?: number | null;
  /** ISO datetime of run */
  date: string;
  /**
     * JSON array of [lat, lng] coordinates
     * @nullable
     */
  route?: string | null;
  /**
     * JSON array of per-km split times in seconds
     * @nullable
     */
  splits?: string | null;
  createdAt: string;
}