# lib/api-zod/src/generated/types/runInput.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Pace Running Tracker API
 * OpenAPI spec version: 0.1.0
 */
import type { RunInputType } from './runInputType';

export interface RunInput {
  name?: string;
  type: RunInputType;
  distance: number;
  duration: number;
  calories?: number;
  elevation?: number;
  date: string;
  /** JSON array of [lat, lng] coordinates */
  route?: string;
  splits?: string;
}