# lib/api-zod/src/generated/types/runUpdate.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Pace Running Tracker API
 * OpenAPI spec version: 0.1.0
 */
import type { RunUpdateType } from './runUpdateType';

export interface RunUpdate {
  name?: string;
  type?: RunUpdateType;
  distance?: number;
  duration?: number;
  calories?: number;
  elevation?: number;
  date?: string;
  route?: string;
  splits?: string;
}