# lib/api-zod/src/generated/types/statsHistory.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Pace Running Tracker API
 * OpenAPI spec version: 0.1.0
 */
import type { DayOfWeekCount } from './dayOfWeekCount';
import type { HistoryPoint } from './historyPoint';

export interface StatsHistory {
  points: HistoryPoint[];
  runsPerDayOfWeek: DayOfWeekCount[];
}