# lib/api-zod/src/generated/types/statsSummary.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Pace Running Tracker API
 * OpenAPI spec version: 0.1.0
 */
import type { AllTimeStats } from './allTimeStats';
import type { DayStatus } from './dayStatus';
import type { PersonalRecords } from './personalRecords';
import type { WeekProgress } from './weekProgress';
import type { WeekStats } from './weekStats';

export interface StatsSummary {
  thisWeek: WeekStats;
  lastWeek: WeekStats;
  allTime: AllTimeStats;
  weekProgress: WeekProgress;
  personalRecords: PersonalRecords;
  weeklyRunDays: DayStatus[];
}