# lib/api-zod/src/generated/types/weekProgress.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Pace Running Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface WeekProgress {
  completed: number;
  goal: number;
  percentage: number;
}